refactor: Moved motion delta calculation out of `data_utils` due to build issues
