+++ src/colorizer/utils/data_utils.ts
@@ -169,129 +169,3 @@ export function formatAsBulletList(items: string[], maxDisplayCount: number = Nu
   }
   return itemDisplayText;
 }
-
-export type TrackData = {
-  ids: number[];
-  times: number[];
-  centroids: number[];
-};
-
-export function makeDataOnlyTracks(
-  trackIds: Uint32Array,
-  times: Uint32Array,
-  centroids: Uint16Array
-): Map<number, TrackData> {
-  const trackIdToTrackData = new Map<number, TrackData>();
-
-  for (let i = 0; i < trackIds.length; i++) {
-    const trackId = trackIds[i];
-    let trackData = trackIdToTrackData.get(trackId);
-    if (!trackData) {
-      trackData = { ids: [], times: [], centroids: [] };
-      trackIdToTrackData.set(trackId, trackData);
-    }
-    trackData.ids.push(i);
-    trackData.times.push(times[i]);
-    trackData.centroids.push(centroids[i * 2], centroids[i * 2 + 1]);
-  }
-
-  // Sort track data by time
-  for (const trackData of trackIdToTrackData.values()) {
-    const indices = [...trackData.times.keys()];
-    indices.sort((a, b) => trackData.times[a] - trackData.times[b]);
-    trackData.times = indices.map((i) => trackData.times[i]);
-    trackData.ids = indices.map((i) => trackData.ids[i]);
-    trackData.centroids = indices.reduce((result, i) => {
-      result.push(trackData.centroids[i * 2], trackData.centroids[i * 2 + 1]);
-      return result;
-    }, [] as number[]);
-  }
-  return trackIdToTrackData;
-}
-
-/**
- * Returns a list of centroid deltas for each object ID in the track. Deltas are calculated
- * as the difference between the centroid at a timepoint and the centroid at the
- * previous timepoint. Up to `numTimesteps` deltas are returned for each object ID.
- * @param track
- * @param numTimesteps
- * @returns
- */
-function getTrackMotionDeltas(track: TrackData, numTimesteps: number): { [key: number]: [number, number][] } {
-  const deltas: { [key: number]: [number, number][] } = {};
-
-  for (let i = 0; i < track.ids.length; i++) {
-    const objectId = track.ids[i];
-    const minTime = track.times[i] - numTimesteps;
-    deltas[objectId] = [];
-
-    for (let j = 0; j < numTimesteps; j++) {
-      // Note that `track.times` is only guaranteed to be ordered, not contiguous.
-      // Check for adjacent, contiguous timepoints that are within the time range.
-      const currentObjectTime = track.times[i - j];
-      const previousObjectTime = track.times[i - j - 1];
-
-      if (currentObjectTime - 1 !== previousObjectTime) {
-        continue;
-      } else if (previousObjectTime < minTime) {
-        break;
-      }
-
-      const currentCentroidX = track.centroids[(i - j) * 2];
-      const currentCentroidY = track.centroids[(i - j) * 2 + 1];
-      const prevCentroidX = track.centroids[(i - j - 1) * 2];
-      const prevCentroidY = track.centroids[(i - j - 1) * 2 + 1];
-      deltas[objectId].push([currentCentroidX - prevCentroidX, currentCentroidY - prevCentroidY]);
-    }
-  }
-  return deltas;
-}
-
-/**
- * Calculates an array of motion deltas for each object in the dataset, averaged over the specified number of timesteps.
- * @param dataset The dataset to calculate motion deltas for.
- * @param numTimesteps The number of timepoints to average over.
- * @param timestepThreshold The minimum number of timepoints the object must have available centroid data for (our of the last
- * `numTimesteps` time points) to be included. For example, if `numTimesteps` is 3 and `timestepThreshold` is 2, the object must
- * exist for at least 2 of the last 3 timepoints to be included in the output.
- * @returns an array of motion deltas, with length equal to `dataset.numObjects * 2`. For each object id `i`, the x and y components
- * of its motion delta are stored at indices `2 * i` and `2 * i + 1`, respectively. If an object does not meet the
- * `timestepThreshold`, both values will be `NaN`.
- */
-export function calculateMotionDeltas(
-  tracks: TrackData[],
-  numTimesteps: number,
-  timestepThreshold: number
-): Float32Array {
-  // Count total objects to allocate the motion deltas array
-  let numObjects = 0;
-  for (const track of tracks) {
-    numObjects += track.ids.length;
-  }
-  const motionDeltas = new Float32Array(numObjects * 2);
-
-  for (const track of tracks) {
-    const objectIdToDeltas = getTrackMotionDeltas(track, numTimesteps);
-
-    for (const [stringObjectId, deltas] of Object.entries(objectIdToDeltas)) {
-      const objectId = parseInt(stringObjectId, 10);
-
-      // Check that the object has enough deltas to meet the threshold; if so
-      // average and store the delta.
-      if (deltas.length >= timestepThreshold) {
-        const averagedDelta: [number, number] = deltas.reduce(
-          (acc, delta) => [acc[0] + delta[0] / deltas.length, acc[1] + delta[1] / deltas.length],
-          [0, 0]
-        );
-        motionDeltas[2 * objectId] = averagedDelta[0];
-        motionDeltas[2 * objectId + 1] = averagedDelta[1];
-      } else {
-        // TODO: These may need to become Infinity for shader compatibility
-        motionDeltas[2 * objectId] = NaN;
-        motionDeltas[2 * objectId + 1] = NaN;
-      }
-    }
-  }
-
-  return motionDeltas;
-}
